requirements.txt
discord.py
google-generativeai
flask